# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PetChain API",
    "version": "1.0.0",
...
    "contact": {
      "name": "PetChain Engineering",
      "email": "engineering@petchain.app",
      "url": "https://petchain.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://staging.petchain.app/api",
      "description": "Staging environment"
    },
    {
      "url": "https://api.petchain.app/api",
      "description": "Production environment"
    }
  ],
...
  "paths": {
...
    "/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "List appointments",
        "description": "Returns appointments. Owners see their pets' appointments; vets see their own appointments; admins see all.",
        "operationId": "listAppointments",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by pet ID"
          },
          {
            "name": "vetId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by vet ID"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW", "RESCHEDULED"]
            },
            "description": "Filter by appointment status"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ROUTINE_CHECKUP",
                "VACCINATION",
                "SURGERY",
                "DENTAL",
                "GROOMING",
                "EMERGENCY",
                "FOLLOW_UP",
                "DIAGNOSTIC",
                "SPECIALIST_REFERRAL",
                "NUTRITION_CONSULTATION"
              ]
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter appointments from this date (inclusive)",
            "example": "2024-03-01"
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter appointments up to this date (inclusive)",
            "example": "2024-03-31"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Appointments"],
        "summary": "Create an appointment",
        "description": "Books a new appointment for a pet with a veterinarian.",
        "operationId": "createAppointment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              },
              "example": {
                "petId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                "vetId": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                "date": "2024-03-15",
                "time": "14:30",
                "durationMinutes": 30,
                "type": "VACCINATION",
                "status": "PENDING",
                "notes": "Annual booster vaccines due",
                "reminder": {
                  "isEnabled": true,
                  "minutesBefore": 60,
                  "notificationMethod": "push"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Appointment"
                    },
                    "message": {
                      "type": "string",
                      "example": "Appointment booked successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Pet or vet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Time slot already booked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "SLOT_UNAVAILABLE",
                    "message": "The requested time slot is already booked"
                  },
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `POST /auth/login` or `POST /auth/register`.\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nTokens expire after 7 days. Use `POST /auth/refresh` to obtain a new token."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server integrations. Contact engineering@petchain.app to obtain a key."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": ["success", "error", "timestamp"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input data"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "example": {
                  "field": "email",
                  "issue": "Invalid format"
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
      "UUID": {
        "type": "string",
        "format": "uuid",
        "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      },
...
      "ISODateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-01-15T10:30:00Z"
      },
...
      "Appointment": {
        "type": "object",
        "required": [
          "id",
          "petId",
          "vetId",
          "date",
          "time",
          "type",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "petId": {
            "$ref": "#/components/schemas/UUID"
          },
          "vetId": {
            "$ref": "#/components/schemas/UUID"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-15",
            "description": "Appointment date in YYYY-MM-DD format"
          },
          "time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "example": "14:30",
            "description": "24-hour time format HH:MM"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 15,
            "maximum": 480,
            "default": 30,
            "example": 30
          },
          "type": {
            "type": "string",
            "enum": [
              "ROUTINE_CHECKUP",
              "VACCINATION",
              "SURGERY",
              "DENTAL",
              "GROOMING",
              "EMERGENCY",
              "FOLLOW_UP",
              "DIAGNOSTIC",
              "SPECIALIST_REFERRAL",
              "NUTRITION_CONSULTATION"
            ],
            "example": "VACCINATION"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW", "RESCHEDULED"],
            "example": "CONFIRMED"
          },
          "notes": {
            "type": "string",
            "example": "Annual booster vaccines due"
          },
          "vet": {
            "$ref": "#/components/schemas/AppointmentVetSummary"
          },
          "pet": {
            "$ref": "#/components/schemas/AppointmentPetSummary"
          },
          "reminder": {
            "$ref": "#/components/schemas/AppointmentReminder"
          },
          "cancelledAt": {
            "$ref": "#/components/schemas/ISODateTime"
          },
          "cancellationReason": {
            "type": "string",
            "example": "Owner unavailable"
          },
          "createdAt": {
            "$ref": "#/components/schemas/ISODateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/ISODateTime"
          }
        }
      },
      "AppointmentVetSummary": {
        "type": "object",
        "required": ["vetId", "name", "clinicName"],
        "properties": {
          "vetId": {
            "$ref": "#/components/schemas/UUID"
          },
          "name": {
            "type": "string",
            "example": "Dr. Sarah Johnson"
          },
          "specialization": {
            "type": "string",
            "example": "General Practice"
          },
          "clinicName": {
            "type": "string",
            "example": "Downtown Animal Hospital"
          },
          "clinicPhone": {
            "type": "string",
            "example": "+14155550100"
          },
          "clinicAddress": {
            "type": "string",
            "example": "123 Main St, San Francisco, CA 94102"
          }
        }
      },
      "AppointmentPetSummary": {
        "type": "object",
        "required": ["petId", "name", "species"],
        "properties": {
          "petId": {
            "$ref": "#/components/schemas/UUID"
          },
          "name": {
            "type": "string",
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "example": "dog"
          },
          "breed": {
            "type": "string",
            "example": "Golden Retriever"
          },
          "age": {
            "type": "number",
            "example": 3.5
          }
        }
      },
      "AppointmentReminder": {
        "type": "object",
        "required": ["isEnabled", "minutesBefore", "notificationMethod"],
        "properties": {
          "isEnabled": {
            "type": "boolean",
            "example": true
          },
          "minutesBefore": {
            "type": "integer",
            "minimum": 5,
            "example": 60,
            "description": "Minutes before appointment to send reminder"
          },
          "notificationMethod": {
            "type": "string",
            "enum": ["push", "email", "sms"],
            "example": "push"
          }
        }
      },
      "CreateAppointmentRequest": {
        "type": "object",
        "required": ["petId", "vetId", "date", "time", "type", "status"],
        "properties": {
          "petId": {
            "$ref": "#/components/schemas/UUID"
          },
          "vetId": {
            "$ref": "#/components/schemas/UUID"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-15"
          },
          "time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "example": "14:30"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 15,
            "maximum": 480,
            "default": 30,
            "example": 30
          },
          "type": {
            "type": "string",
            "enum": [
              "ROUTINE_CHECKUP",
              "VACCINATION",
              "SURGERY",
              "DENTAL",
              "GROOMING",
              "EMERGENCY",
              "FOLLOW_UP",
              "DIAGNOSTIC",
              "SPECIALIST_REFERRAL",
              "NUTRITION_CONSULTATION"
            ],
            "example": "VACCINATION"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW", "RESCHEDULED"],
            "default": "PENDING",
            "example": "PENDING"
          },
          "notes": {
            "type": "string",
            "example": "Annual booster vaccines due"
          },
          "reminder": {
            "$ref": "#/components/schemas/AppointmentReminder"
          }
        }
      },
...
      "AppointmentListResponse": {
        "type": "object",
        "required": ["success", "data", "total"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "total": {
            "type": "integer",
            "example": 5
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "example": 20
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Number of items per page"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — invalid input or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid input data",
                "details": {
                  "field": "email",
                  "issue": "Invalid email format"
                }
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required. Please log in."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "Server error. Please try again later."
              },
              "timestamp": "2024-01-15T10:30:00Z"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}